made changes

+++ JavaImplementation/src/org/learn/interview/Point.java
@@ -1,10 +1,10 @@
 package org.learn.interview;
 
 public class Point {
-	public Integer x;
-	public Integer y;
+	public float x;
+	public float y;
 
-	public Point(int x, int y) {
+	public Point(float x, float y) {
 		this.x = x;
 		this.y = y;
 	}
